+++ packages/blinkid-verify-ux-manager/src/ui/BlinkIdVerifyFeedbackUi.tsx
@@ -154,6 +154,14 @@ export const BlinkIdVerifyFeedbackUi: Component<{
     }
   });
 
+  createEffect(() => {
+    if (displayUnsupportedDocumentModal()) {
+      void store.blinkIdVerifyUxManager.analytics.logAlertDisplayedEvent(
+        "DocumentNotSupported",
+      );
+    }
+  });
+
   return (
     <div>
       <style