+++ packages/blinkid-worker/src/BlinkIdWorker.test.ts
@@ -55,7 +55,7 @@ describe("BlinkIdWorker", () => {
     const worker = new BlinkIdWorker();
     const pinglet: Ping = {
       schemaName: "ping.sdk.init.start",
-      schemaVersion: "1.1.0",
+      schemaVersion: "1.2.0",
       sessionNumber: 0,
       data: {
         product: "BlinkID",