+++ packages/blinkid-verify-worker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @microblink/blinkid-verify-worker
 
+## 3.20.1
+
+### Patch Changes
+
+- Bumped tracked schema version for the `ping.sdk.init.start`
+- Updated dependencies
+  - @microblink/analytics@2.0.0
+  - @microblink/blinkid-verify-wasm@3.20.1
+
 ## 3.20.0
 
 - Introducing BlinkID Verify web SDK, a capturing solution for perparing the perfect frames from a camera to be sent to the BlinkID verify API