+++ packages/blinkid-verify/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @microblink/blinkid-verify
 
+## 3.20.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @microblink/blinkid-verify-ux-manager@3.20.1
+  - @microblink/blinkid-verify-core@3.20.1
+
 ## 3.20.0
 
 - Introducing BlinkID Verify web SDK, a capturing solution for perparing the perfect frames from a camera to be sent to the BlinkID verify API